django_backend/requirements.txt
Django==4.2.11
celery==5.3.6
psycopg2-binary==2.9.9
redis==4.3.4
pytest
pytest-django
pytest-cov
factory-boy
python-dotenv
djangorestframework
djangorestframework-simplejwt	